@@ -1,10 +1,10 @@
 name: chrome_launcher
 description: Android Flutter Plugin to force launch an URL with Google Chrome.
-version: 1.1.0
+version: 1.2.0
 homepage: https://github.com/nsNeruno/
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
